requirements.txt
sounddevice
numpy
scipy
openwakeword
onnxruntime
anthropic
elevenlabs
duckduckgo-search
Pillow
python-dotenv
